+++ src/extension.ts
@@ -4,20 +4,21 @@ import {
   applyShellEdit,
   applyShellFilter,
 } from './applyShellEdit'
-import { register as registerCaseTransform } from './caseTransform'
 import { setExtContext } from './context'
 import { register as registerCssSeletorComplete } from './cssSelectorComplete'
 import { PathCompleteProvider } from './pathComplete'
 import { quicklySwitchFile } from './quicklySwitchFile'
+import { renameWithCase, transformCase } from './transformCase'
 import { register as registerTsCodeAction } from './tsCodeActions/selection'
 
 export function activate(context: ExtensionContext) {
   setExtContext(context)
-  PathCompleteProvider.register!()
   registerCssSeletorComplete()
-  registerCaseTransform()
   registerTsCodeAction()
   context.subscriptions.push(
+    new PathCompleteProvider(),
+    commands.registerCommand('mvext.renameWithCase', renameWithCase),
+    commands.registerTextEditorCommand('mvext.transformCase', transformCase),
     commands.registerCommand('mvext.quicklySwitchFile', quicklySwitchFile),
     commands.registerCommand('mvext.applyShellEdit', applyShellEdit),
   )