+++ CHANGELOG.md
@@ -8,6 +8,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [0.3.0]
 
-- Feature: [`caseTransform`](./src/caseTransform.ts): transform word selections to camelCase, pascalCase, snakeCase, kebabCase, titleCase, headerCase, dotCase, pathCase, noCase, lowerCase, upperCase.
+- Feature: [`transformCase`](./src/transformCase.ts): transform word selections to camel, pascal, snake, kebab, title, header, dot, path, normal, lower, upper.
 - Feature: [`quicklySwitchFile`](./src/quicklySwitchFile.ts): switch files between html and css/js or switch tests files.
 - Feature: [`applyShellEdit`](./src/applyShellEdit.ts): exec selection in scripting languages and replace selection by results, except empty or null.